requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
litellm>=1.30.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
aiosmtplib>=3.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
email-validator>=2.1.0
